z8000: fix size_t to 32-bit, exec reaches /init
size_t was 16-bit but the malloc allocator uses bit 31 of the block
length as the USED flag (UNUSED macro checks 1L<<31). With 16-bit
size_t, this check always returned true (bit 31 of a 16-bit value
is always 0), causing kfree to panic with "bad free" on every free.

Changed size_t to unsigned long (32-bit) matching the 68000 port.

Also increased UBLOCK_SIZE to 2048 for Z8001's deeper stack usage
with 32-bit pointers and arg_t.

The kernel now successfully finds /init, passes exec validation,
and attempts to execute it. The execve returns with err=0, which
likely means exec succeeded and the user program ran but then
exited (sc #0 = _exit), returning control to the kernel with no
other processes to schedule.

Next: implement proper /init that loops, or handle the _exit case.

+++ Kernel/cpu-z8000/types.h
@@ -4,7 +4,7 @@ typedef unsigned short uint16_t;
 typedef signed short int16_t;
 typedef unsigned char uint8_t;
 typedef signed char int8_t;
-typedef unsigned short size_t;
+typedef unsigned long size_t;		/* Must be 32-bit: malloc uses bit 31 as USED flag */
 
 typedef unsigned char uint_fast8_t;
 typedef signed char int_fast8_t;
